Advanced logic (#38)
* Created components

* Created config editor

* Created analysis options component

* Added defaults to anlaysis options

* Fixed e2e tests

* Attempt to fix azure-pipelines tests

* Regrouped advanced components

* Working analysis options editor

* Working reset buttons for advanced analysis

* Pontentially working advanced analysis

* Linted project

* Working advanced analysis options

Currently there is an issue where the original config and options are being overwritten in a way which cannot be reverted for later analyses

* Separate class files

* Cleanup class names

* Working config editor

Changed how advanced component works. Now instead of parsing the currentAnalysis, it takes an input of option and config

* Passing build, analysis broken

Fixed the software to allow the azure pipelines build to pass. However functionality of the application is currently in a broken state.